requirements.txt
Django==5.0.6
djangorestframework==3.15.0
mysqlclient==2.2.4
django-mysql==4.13.0
python-dotenv
service_registrar @ git+https://github.com/kodev8/healthcare-service-registrar.git@8c3dcf8b4fcb364bfdde740984b2b3042f9385fd